minifer: Split pass
