+++ ecmascript/minifier/src/metadata/mod.rs
@@ -42,16 +42,6 @@ struct InfoMarker<'a> {
 }
 
 impl InfoMarker<'_> {
-    fn make_unique(&self, span: &mut Span) {
-        debug_assert_eq!(
-            span.ctxt,
-            SyntaxContext::empty(),
-            "Expected empty syntax context"
-        );
-
-        span.ctxt = span.ctxt.apply_mark(Mark::fresh(Mark::root()));
-    }
-
     /// Check for `/** @const */`.
     pub(super) fn has_const_ann(&self, span: Span) -> bool {
         self.find_comment(span, |c| {
@@ -116,12 +106,6 @@ impl InfoMarker<'_> {
 impl VisitMut for InfoMarker<'_> {
     noop_visit_mut_type!();
 
-    fn visit_mut_block_stmt(&mut self, n: &mut BlockStmt) {
-        n.visit_mut_children_with(self);
-
-        self.make_unique(&mut n.span);
-    }
-
     fn visit_mut_call_expr(&mut self, n: &mut CallExpr) {
         n.visit_mut_children_with(self);
 
@@ -173,12 +157,6 @@ impl VisitMut for InfoMarker<'_> {
                 return;
             }
         }
-    }
-
-    fn visit_mut_function(&mut self, n: &mut Function) {
-        n.visit_mut_children_with(self);
-
-        self.make_unique(&mut n.span);
     }
 
     fn visit_mut_ident(&mut self, _: &mut Ident) {}
