+++ ecmascript/minifier/src/pass/mod.rs
@@ -6,3 +6,4 @@ pub mod mangle_names;
 pub mod mangle_props;
 pub mod postcompress;
 pub mod precompress;
+pub mod unique_scope;
